requirements.txt
# Core dependencies
streamlit
nest-asyncio==1.6.0

# Google Cloud & AI
google-genai==1.40.0
google-generativeai
google-cloud-discoveryengine
google-cloud-secret-manager==2.10.0
google-api-python-client==2.182.0
google-auth==2.41.0
google-crc32c==1.3.0

# Data processing
numpy
scikit-learn
pandas

# Web scraping
playwright
requests

# Utilities
python-dotenv
gspread
gspread-dataframe
